Texture Quality Adjustments
r.VT.MaxUploadsPerFrame seems to be modified by the game. Not sure why but sometimes it can get stuck at the lowest possible value (say TextureQuality@0 = 8, it will be 8).

I'm setting it as a baseline of 32, and doubling from there for each Texture Quality setting just to be in the safe side.